# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Still Afloat Editorial Agent API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  success: boolean;
  error: string;
}

export interface ActionResult {
  success: boolean;
  message?: string;
}

export interface Story {
  id: string;
  title: string;
  /** @nullable */
  category?: string | null;
  /** @nullable */
  impactLevel?: string | null;
  /** @nullable */
  travelerImpact?: string | null;
  /** @nullable */
  summary?: string | null;
  /** @nullable */
  reasoning?: string | null;
  /** @nullable */
  link?: string | null;
  /** @nullable */
  image?: string | null;
  /** @nullable */
  featured?: boolean | null;
  /** @nullable */
  homepageCandidate?: boolean | null;
  sources?: string[];
  sourceLinks?: string[];
  /** @nullable */
  publishedAt?: string | null;
  /** @nullable */
  approvedAt?: string | null;
  /** @nullable */
  status?: string | null;
}

export interface Alert {
  id: string;
  title: string;
  /** @nullable */
  category?: string | null;
  /** @nullable */
  impactLevel?: string | null;
  /** @nullable */
  travelerImpact?: string | null;
  /** @nullable */
  summary?: string | null;
  /** @nullable */
  link?: string | null;
  /** @nullable */
  approvedAt?: string | null;
  /** @nullable */
  featured?: boolean | null;
}

export interface EditorialQueueResponse {
  success: boolean;
  /** @nullable */
  generatedAt?: string | null;
  degradedMode?: boolean;
  count: number;
  stories: Story[];
}

export interface StoryFeedResponse {
  success: boolean;
  /** @nullable */
  source?: string | null;
  /** @nullable */
  generatedAt?: string | null;
  count: number;
  stories: Story[];
}

export interface AlertsFeedResponse {
  success: boolean;
  /** @nullable */
  source?: string | null;
  /** @nullable */
  generatedAt?: string | null;
  count: number;
  alerts: Alert[];
}

export interface WeatherAlertsResponse {
  success: boolean;
  /** @nullable */
  source?: string | null;
  /** @nullable */
  generatedAt?: string | null;
  monitoringEnabled?: boolean;
  monitoredPorts: string[];
  status?: string;
}

export type SystemStatusSystems = {
  openaiConfigured?: boolean;
  resendConfigured?: boolean;
  gnewsConfigured?: boolean;
  weatherConfigured?: boolean;
  approvalConfigured?: boolean;
};

export type SystemStatusPublishing = {
  candidateStories?: number;
  approvedStories?: number;
  homepageStories?: number;
};

export type SystemStatusPipeline = {
  degradedMode?: boolean;
  /** @nullable */
  degradedReason?: string | null;
};

export interface SystemStatus {
  success: boolean;
  service: string;
  generatedAt: string;
  environment?: string;
  systems?: SystemStatusSystems;
  publishing?: SystemStatusPublishing;
  pipeline?: SystemStatusPipeline;
}

export type ScanResultTelemetry = { [key: string]: unknown };

export type ScanResultEmailResult = { [key: string]: unknown };

export interface ScanResult {
  success: boolean;
  scannedStories?: number;
  curatedStories?: number;
  degradedMode?: boolean;
  telemetry?: ScanResultTelemetry;
  emailResult?: ScanResultEmailResult;
}

export type PlatformManifestEndpoints = { [key: string]: unknown };

export type PlatformManifestCapabilities = { [key: string]: unknown };

export interface PlatformManifest {
  success: boolean;
  service: string;
  generatedAt: string;
  endpoints?: PlatformManifestEndpoints;
  capabilities?: PlatformManifestCapabilities;
}

export type PerformAgentActionParams = {
action: PerformAgentActionAction;
id: string;
token?: string;
};

export type PerformAgentActionAction = typeof PerformAgentActionAction[keyof typeof PerformAgentActionAction];


export const PerformAgentActionAction = {
  approve: 'approve',
  reject: 'reject',
  pin: 'pin',
  defer: 'defer',
} as const;

export type GetStoryDetailsParams = {
id?: string;
};