# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Still Afloat Editorial Agent API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /agent-action:
    get:
      operationId: performAgentAction
      tags: [editorial]
      summary: Perform editorial action on a story
      description: Approve, reject, pin, or defer a candidate story
      parameters:
        - name: action
          in: query
          required: true
          schema:
            type: string
            enum: [approve, reject, pin, defer]
        - name: id
          in: query
          required: true
          schema:
            type: string
        - name: token
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Action result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActionResult"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Story not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /story-details:
    get:
      operationId: getStoryDetails
      tags: [feeds]
      summary: Get story details
      description: Returns full story details, optionally filtered by ID
      parameters:
        - name: id
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Story details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StoryFeedResponse"
        "404":
          description: Story not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /scan-news:
    post:
      operationId: scanNews
      tags: [agent]
      summary: Trigger a news scan
      description: Fetches live news, runs AI curation, stores candidates, sends digest email
      responses:
        "200":
          description: Scan result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScanResult"
        "500":
          description: Scan failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
        error:
          type: string
      required:
        - success
        - error
    ActionResult:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
    Story:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        category:
          type: ["string", "null"]
        impactLevel:
          type: ["string", "null"]
        travelerImpact:
          type: ["string", "null"]
        summary:
          type: ["string", "null"]
        reasoning:
          type: ["string", "null"]
        link:
          type: ["string", "null"]
        image:
          type: ["string", "null"]
        featured:
          type: ["boolean", "null"]
        homepageCandidate:
          type: ["boolean", "null"]
        sources:
          type: array
          items:
            type: string
        sourceLinks:
          type: array
          items:
            type: string
        publishedAt:
          type: ["string", "null"]
        approvedAt:
          type: ["string", "null"]
        status:
          type: ["string", "null"]
      required:
        - id
        - title
...
    StoryFeedResponse:
      type: object
      properties:
        success:
          type: boolean
        source:
          type: ["string", "null"]
        generatedAt:
          type: ["string", "null"]
        count:
          type: integer
        stories:
          type: array
          items:
            $ref: "#/components/schemas/Story"
      required:
        - success
        - count
        - stories
...
    ScanResult:
      type: object
      properties:
        success:
          type: boolean
        scannedStories:
          type: integer
        curatedStories:
          type: integer
        degradedMode:
          type: boolean
        telemetry:
          type: object
        emailResult:
          type: object
      required:
        - success
...