# lib/api-zod/src/generated/types/actionResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Still Afloat Editorial Agent API
 * OpenAPI spec version: 0.1.0
 */

export interface ActionResult {
  success: boolean;
  message?: string;
}