# lib/api-zod/src/generated/types/story.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Still Afloat Editorial Agent API
 * OpenAPI spec version: 0.1.0
 */

export interface Story {
  id: string;
  title: string;
  /** @nullable */
  category?: string | null;
  /** @nullable */
  impactLevel?: string | null;
  /** @nullable */
  travelerImpact?: string | null;
  /** @nullable */
  summary?: string | null;
  /** @nullable */
  reasoning?: string | null;
  /** @nullable */
  link?: string | null;
  /** @nullable */
  image?: string | null;
  /** @nullable */
  featured?: boolean | null;
  /** @nullable */
  homepageCandidate?: boolean | null;
  sources?: string[];
  sourceLinks?: string[];
  /** @nullable */
  publishedAt?: string | null;
  /** @nullable */
  approvedAt?: string | null;
  /** @nullable */
  status?: string | null;
}