# lib/api-zod/src/generated/types/storyFeedResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Still Afloat Editorial Agent API
 * OpenAPI spec version: 0.1.0
 */
import type { Story } from './story';

export interface StoryFeedResponse {
  success: boolean;
  /** @nullable */
  source?: string | null;
  /** @nullable */
  generatedAt?: string | null;
  count: number;
  stories: Story[];
}